Womensscripttransliterator 27 (#32)
* Resolves #27

Fixed error where units which spanned atoms would cause the subatom for the part of the unit which extended to the second atom to have the context of the first atom, rather than the second. Because the reducers at the end of the transliteration step use the context of the first subatom for a given atom as the context for all subatoms of that transliterated atom, the second atom was not able to have its transliteration inserted into the xml document (the replacement of the contents of the first atom made that target invalid).

* Resolves #19

Changed words in dictionary to use a literal zero-width joiner rather than a unicode escape sequence.

* Resolves #12

Rewrote the StringTransliteror.transliterateAtoms() function to have logic that consumes orphaned bits of a subunit in the next unit's atom.
Removed StringUnit.matchesEndPattern() and replaced uses of it with StringUnit.splitIntoSubunits().
Removed StringUnit.isComplete field, since it wasn't actually used for anything anymore.